+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed badges.
 - Added back configurations and dependencies.
 - Fixing issue with Cluster when only NodeMajority is used.
+  - Fixed the fix: Quorum is not required in some SQL Always-On scenarios which did not work after the fix.
 - Add new resource LocalUsers.
 - Make DscLcmController independent from the DscDiagnostics resource.
 - Add optional attributes to DscTagging resource.
@@ -24,3 +25,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added MmaAgent to configure Microsoft Monitoring Agent.
 - Added AddsServicePrincipalNames to configure SPNs.
 - Disabling RebootNodeIfNeeded when LCM is on Monitor mode.
+- Made 'WaitForClusterRetryIntervalSec' and 'WaitForClusterRetryCount' configurable in Cluster config.